+++ src/components/FilePreview.tsx
@@ -1,9 +1,5 @@
 import React from 'react';
 
-const FilePreview: React.FC<any> = ({ onClick, children, style }) => (
-    <div className="filePreview" onClick={onClick} style={style}>
-        {children}
-    </div>
-);
+const FilePreview: React.FC<any> = () => <div className="file-preview"></div>;
 
 export default FilePreview;