+++ src/containers/Home.tsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { HomeState } from './models/state';
 
 import SideMenu from '../components/SideMenu';
+import FileViewer from '../components/FileViewer';
 
 class Home extends Component {
     state: HomeState = {
@@ -11,7 +12,12 @@ class Home extends Component {
         const { sideMenuOpen } = this.state;
         const menuView = sideMenuOpen ? <SideMenu /> : null;
 
-        return <div id="home">{menuView}</div>;
+        return (
+            <div id="home">
+                {menuView}
+                <FileViewer />
+            </div>
+        );
     }
 }